fix: api의 다른 기능으로 인해 발생하는 circular ref 이슈 조치를 위한 ping 기능 분리
