+++ src/utils/api.ts
@@ -275,40 +275,3 @@ export function showModal(options: ModalOptions): void {
   overlay.appendChild(modal);
   document.body.appendChild(overlay);
 }
-
-export function ping(retryCount?:number): Promise<void> {
-  if (!retryCount) {
-    retryCount = 3;
-  } else if (retryCount <= 0) {
-    return Promise.reject(new Error("Ping failed: Maximum retry attempts reached"));
-  }
-
-  return new Promise<void>((resolve, reject) => {
-    try {
-      const port = chrome.runtime.connect({ name: 'ping' });
-
-      port.onMessage.addListener((msg) => {
-        if (msg.type === 'PONG') {
-          port.disconnect();
-          console.info("Connection successful");
-          resolve();
-        }
-      });
-
-      port.onDisconnect.addListener(() => {
-        if (chrome.runtime.lastError) {
-          // reject(new Error("Ping failed: " + chrome.runtime.lastError.message));
-          console.info("Ping failed, retrying...");
-          setTimeout(() => {
-            ping(retryCount! - 1).then(resolve).catch(reject);
-          }, 500);
-        }
-      });
-    } catch (error) {
-      console.error("Ping error:", error);
-      setTimeout(() => {
-        ping(retryCount! - 1).then(resolve).catch(reject);
-      }, 500);
-    }
-  });
-}
